RAG_YouTube_Project/requirements.txt
youtube-transcript-api
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain-google-genai
langchain_community
langchain_core